requirements.txt
pydantic
python-dotenv
tiktoken
openai
streamlit
streamlit-audiorec
pydub
sympy
requests
srt
PyPDF2
diskcache
aiohttp
aiofiles
markdown==3.5.1
redis==5.1.1
pyjwt
azure-ai-formrecognizer
azure-storage-blob
azure-ai-translation-document
azure-ai-translation-text
azure-cognitiveservices-speech
azure-search-documents
llama-index
streamlit-paste-button
streamlit-ace
psycopg2-binary
sqlalchemy
wordcloud